fixing navbar refactor
