+++ src/App.tsx
@@ -5,12 +5,10 @@ import { AosInitializer } from "./lib/aosInitializer";
 
 const App = () => {
   return (
-    <div className="dark:bg-black">
-      <Layout>
-        <AosInitializer />
-        <Outlet />
-      </Layout>
-    </div>
+    <Layout>
+      <AosInitializer />
+      <Outlet />
+    </Layout>
   );
 };
