+++ src/components/page-container/PageContainer.tsx
@@ -13,7 +13,10 @@ export const PageContainer = ({
   className,
 }: PageContainerProps) => {
   return (
-    <section className={cn("h-full px-10 py-16", className)} id={id}>
+    <section
+      className={cn("container mx-auto h-full px-10 py-16", className)}
+      id={id}
+    >
       {children}
     </section>
   );
