+++ src/main.tsx
@@ -5,7 +5,7 @@ import { RouterProvider } from "react-router-dom";
 import "virtual:svg-icons-register";
 import "./index.css";
 import appRouter from "./router/appRouter.tsx";
-import { ThemeProvider } from "./components/theme-provider/ThemeProvider.tsx";
+import { ThemeProvider } from "./context/theme-provider/ThemeProvider.tsx";
 
 import "@/i18n/index.ts";
